gradients: nicer gradients.
The colours on the markers of the map are now calculated as
interpolated values in the HSV colour space.

There is also an implementation for using a heat map
to display data, but it does not really work and can
unfortunately not be used (it shows wrong values but would
be much faster).